+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -187,7 +187,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KYQ73NJYBCGTTJVTV9683KMW"),
+  _id: z.string().default("01KYRFKV0NDFK6WA55WAKM6AD1"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: RetrieveDatasourceStatus$inboundSchema,