+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-29T15:15:09.613Z",
+    "2026-07-30T03:03:02.613Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {